refactor(rexglue): drop config migration and legacy conversion
