+++ src/rexglue/CMakeLists.txt
@@ -7,8 +7,7 @@ add_executable(rexglue
     commands/codegen_command.cpp
     commands/init_command.cpp
     commands/test_recompiler.cpp
-    commands/migration_scan.cpp
-    commands/legacy_config.cpp
+    commands/project_scan.cpp
 )
 
 rex_config_postfix(rexglue)
