+++ src/rexglue/commands/codegen_command.cpp
@@ -9,11 +9,9 @@
 #include "codegen_command.h"
 #include "../ui/progress.h"
 #include "../ui/ui.h"
-#include "legacy_config.h"
-#include "migration_scan.h"
+#include "project_scan.h"
 #include "template_utils.h"
 
-#include <algorithm>
 #include <filesystem>
 #include <memory>
 #include <span>
@@ -23,7 +21,6 @@
 
 #include <CLI/CLI.hpp>
 #include <fmt/format.h>
-#include <toml++/toml.hpp>
 
 #include <rex/codegen/manifest.h>
 #include <rex/codegen/project_recompiler.h>
@@ -36,122 +33,7 @@ namespace {
 
 namespace fs = std::filesystem;
 
-// Compare dot-separated numeric version strings (e.g. "0.7.5"). Trailing
-// pre-release/build metadata after '-' or '+' is ignored. Returns true when
-// lhs orders strictly before rhs.
-bool VersionLess(std::string_view lhs, std::string_view rhs) {
-  auto parse = [](std::string_view s) {
-    std::vector<unsigned> parts;
-    unsigned value = 0;
-    bool in_number = false;
-    for (char c : s) {
-      if (c >= '0' && c <= '9') {
-        value = value * 10 + static_cast<unsigned>(c - '0');
-        in_number = true;
-        continue;
-      }
-      if (in_number)
-        parts.push_back(value);
-      value = 0;
-      in_number = false;
-      if (c == '-' || c == '+')
-        break;
-    }
-    if (in_number)
-      parts.push_back(value);
-    return parts;
-  };
-  auto a = parse(lhs);
-  auto b = parse(rhs);
-  for (std::size_t i = 0; i < std::max(a.size(), b.size()); ++i) {
-    unsigned va = i < a.size() ? a[i] : 0;
-    unsigned vb = i < b.size() ? b[i] : 0;
-    if (va != vb)
-      return va < vb;
-  }
-  return false;
-}
-
-// The upgrade scan only runs when the installed SDK is newer than the version
-// that last generated this project (or when the project was never stamped).
-bool MigrationNeeded(std::string_view last_codegen_version, std::string_view installed_version) {
-  return last_codegen_version.empty() || VersionLess(last_codegen_version, installed_version);
-}
-
-struct ActionStrings {
-  std::string_view verb;
-  std::string_view label;
-};
-
-ActionStrings ActionStringsFor(OverwriteAction action) {
-  switch (action) {
-    case OverwriteAction::Write:
-      return {"Wrote", "write "};
-    case OverwriteAction::Delete:
-      return {"Deleted", "delete"};
-  }
-  return {"Touched", "?     "};
-}
-
-bool ApplyEntry(const OverwriteEntry& entry) {
-  std::error_code ec;
-  switch (entry.action) {
-    case OverwriteAction::Write: {
-      if (auto parent = entry.path.parent_path(); !parent.empty()) {
-        fs::create_directories(parent, ec);
-        if (ec) {
-          REXLOG_ERROR("Failed to create directory for {}: {}", entry.path.string(), ec.message());
-          return false;
-        }
-      }
-      return write_file_atomic(entry.path, entry.rendered_content);
-    }
-    case OverwriteAction::Delete:
-      if (!fs::exists(entry.path))
-        return true;
-      if (!fs::remove(entry.path, ec) || ec) {
-        REXLOG_ERROR("Failed to delete {}: {}", entry.path.string(), ec.message());
-        return false;
-      }
-      return true;
-  }
-  return false;
-}
-
-Result<void> ApplyPlan(const std::vector<OverwriteEntry>& plan) {
-  for (const auto& entry : plan) {
-    if (!ApplyEntry(entry)) {
-      return Err<void>(rex::ErrorCategory::IO,
-                       fmt::format("Failed to apply: {}", entry.path.string()));
-    }
-    REXLOG_TRACE("{}: {}", ActionStringsFor(entry.action).verb, entry.path.generic_string());
-  }
-  return rex::Ok();
-}
-
-Result<void> PromptConsent(const std::vector<OverwriteEntry>& plan, bool force) {
-  std::vector<ui::PlanRow> rows;
-  for (const auto& e : plan) {
-    if (e.silent)
-      continue;
-    rows.push_back({ActionStringsFor(e.action).label, e.path.generic_string(), e.reason});
-  }
-  if (rows.empty())
-    return rex::Ok();
-
-  ui::PlanTable(
-      fmt::format("Migration: {} file(s) will be rewritten before codegen runs.", rows.size()),
-      rows);
-
-  if (force)
-    return rex::Ok();
-  if (!ui::Confirm("Apply migration and continue?")) {
-    return Err<void>(rex::ErrorCategory::UserAbort, "Upgrade declined; codegen aborted.");
-  }
-  return rex::Ok();
-}
-
-void EmitManualReview(std::span<const MigrationWarning> warnings, std::string_view header) {
+void EmitManualReview(std::span<const SourceWarning> warnings, std::string_view header) {
   if (warnings.empty())
     return;
   std::vector<ui::ManualReviewRow> rows;
@@ -161,28 +43,6 @@ void EmitManualReview(std::span<const MigrationWarning> warnings, std::string_vi
         {fmt::format("{}:{}", w.file.generic_string(), w.line_number), w.detail, w.hint});
   }
   ui::ManualReviewList(std::string(header), rows);
-}
-
-MigrationFindings ScanProjectMigrations(const fs::path& project_dir, std::string_view project_name,
-                                        std::string_view sdk_version,
-                                        std::string_view entrypoint_out_dir) {
-  MigrationFindings out;
-  auto add_rewrites = [&](std::vector<OverwriteEntry> entries) {
-    out.rewrites.insert(out.rewrites.end(), std::make_move_iterator(entries.begin()),
-                        std::make_move_iterator(entries.end()));
-  };
-  auto add_warnings = [&](std::vector<MigrationWarning> entries) {
-    out.warnings.insert(out.warnings.end(), std::make_move_iterator(entries.begin()),
-                        std::make_move_iterator(entries.end()));
-  };
-
-  add_rewrites(ScanSdkTemplateDrift(project_dir, project_name, sdk_version, entrypoint_out_dir));
-  add_rewrites(ScanSourceIncludeRewrites(project_dir, project_name));
-  auto idents = ScanLegacyIdentifiers(project_dir);
-  add_rewrites(std::move(idents.rewrites));
-  add_warnings(std::move(idents.warnings));
-  add_warnings(ScanCallSitePatterns(project_dir));
-  return out;
 }
 
 void ReportStaleIncludes(const fs::path& manifest_path,
@@ -205,7 +65,7 @@ void ReportStaleIncludes(const fs::path& manifest_path,
 }
 
 Result<void> RecompileProject(const fs::path& manifest_path, const CliContext& ctx,
-                              const std::vector<std::string>& targets) {
+                              const std::vector<std::string>& targets, bool ignore_stamp) {
   auto manifest = rex::codegen::ManifestConfig::Load(manifest_path);
   if (!manifest) {
     return Err<void>(rex::ErrorCategory::Config, "Failed to load manifest");
@@ -217,12 +77,17 @@ Result<void> RecompileProject(const fs::path& manifest_path, const CliContext& c
       .targets = targets,
       .force = ctx.generate_despite_errors,
       .reporter = &progress,
+      .ignoreStamp = ignore_stamp,
+      .sdkVersion = REXGLUE_VERSION_FLOOR "-" REXGLUE_VERSION_CHANNEL,
   };
   auto result = recompiler.Run(opts);
   if (!result)
     return result;
 
   ReportStaleIncludes(manifest_path, recompiler);
+  REXLOG_INFO("Codegen summary: {} written, {} unchanged, {} deleted, {} module(s) up to date",
+              recompiler.writtenFiles().size(), recompiler.unchangedFiles().size(),
+              recompiler.deletedFiles().size(), recompiler.skippedModules().size());
   return rex::Ok();
 }
 
@@ -268,6 +133,7 @@ void EmitProjectHeader(const fs::path& manifest_path, const ManifestSummary& sum
 struct CodegenArgs {
   std::string config_path;
   std::vector<std::string> targets;
+  bool ignore_stamp = false;
 };
 
 }  // namespace
@@ -316,118 +182,37 @@ Result<std::string> DiscoverManifestInCwd() {
 }
 
 Result<void> CodegenFromConfig(const std::string& config_path, const CliContext& ctx,
-                               const std::vector<std::string>& targets) {
+                               const std::vector<std::string>& targets, bool ignore_stamp) {
   REXLOG_TRACE("Generating code with config: {}", config_path);
 
-  toml::table parsed_tbl;
-  try {
-    parsed_tbl = toml::parse_file(config_path);
-  } catch (const toml::parse_error& err) {
-    return Err<void>(rex::ErrorCategory::Config,
-                     fmt::format("Failed to parse {}: {}", config_path, err.what()));
+  fs::path manifest_path = config_path;
+  if (!rex::codegen::ManifestConfig::IsManifest(manifest_path)) {
+    return Err<void>(
+        rex::ErrorCategory::Config,
+        fmt::format("{} is not a manifest: no [project] section. Pre-manifest single-file "
+                    "configs are no longer converted automatically; see docs for the layout.",
+                    config_path));
   }
 
-  std::vector<OverwriteEntry> pre_plan;
-  std::vector<OverwriteEntry> post_plan;
-  std::vector<MigrationWarning> warnings;
-  bool from_legacy = false;
+  auto loaded = LoadManifestSummary(manifest_path);
+  if (!loaded)
+    return Err<void>(loaded.error());
+  auto summary = std::move(*loaded);
 
-  fs::path manifest_path;
-  ManifestSummary summary;
   const std::string current_version = REXGLUE_VERSION_FLOOR;
+  EmitProjectHeader(manifest_path, summary);
 
-  auto append_findings = [&](MigrationFindings findings) {
-    post_plan.insert(post_plan.end(), std::make_move_iterator(findings.rewrites.begin()),
-                     std::make_move_iterator(findings.rewrites.end()));
-    warnings.insert(warnings.end(), std::make_move_iterator(findings.warnings.begin()),
-                    std::make_move_iterator(findings.warnings.end()));
-  };
-
-  if (parsed_tbl.contains("project")) {
-    manifest_path = config_path;
-    auto loaded = LoadManifestSummary(manifest_path);
-    if (!loaded)
-      return Err<void>(loaded.error());
-    summary = std::move(*loaded);
-    if (MigrationNeeded(summary.sdk_version, current_version)) {
-      append_findings(ScanProjectMigrations(manifest_path.parent_path(), summary.project_name,
-                                            current_version, summary.entrypoint_out_dir));
-    }
-  } else {
-    fs::path legacy_path = config_path;
-    auto converted = ConvertLegacyConfig(legacy_path);
-    if (!converted) {
-      return Err<void>(
-          rex::ErrorCategory::Config,
-          fmt::format("Cannot convert legacy config {}: missing project_name or file_path",
-                      legacy_path.string()));
-    }
-    manifest_path = converted->manifest_path;
-    summary.project_name = converted->project_name;
-    summary.entrypoint_out_dir = converted->out_directory_path;
-
-    if (fs::exists(manifest_path)) {
-      REXLOG_WARN("Both {} and {} exist; using the manifest. Remove the legacy file when ready.",
-                  manifest_path.filename().string(), legacy_path.filename().string());
-      auto loaded = LoadManifestSummary(manifest_path);
-      if (!loaded)
-        return Err<void>(loaded.error());
-      summary = std::move(*loaded);
-      if (MigrationNeeded(summary.sdk_version, current_version)) {
-        append_findings(ScanProjectMigrations(manifest_path.parent_path(), summary.project_name,
-                                              current_version, summary.entrypoint_out_dir));
-      }
-    } else {
-      from_legacy = true;
-      pre_plan.push_back({manifest_path, std::move(converted->manifest_content),
-                          OverwriteAction::Write, /*silent=*/false,
-                          fmt::format("upgrade format to v{}", current_version)});
-      if (!converted->stripped_legacy_content.empty()) {
-        post_plan.push_back({legacy_path, std::move(converted->stripped_legacy_content),
-                             OverwriteAction::Write, /*silent=*/false,
-                             "strip absorbed fields (kept as include target)"});
-      } else {
-        post_plan.push_back({legacy_path, "", OverwriteAction::Delete, /*silent=*/false,
-                             "absorbed into the new manifest"});
-      }
-
-      auto cmake_rewrites =
-          ScanCmakeReferences(legacy_path.parent_path(), legacy_path.filename().string(),
-                              manifest_path.filename().string());
-      post_plan.insert(post_plan.end(), std::make_move_iterator(cmake_rewrites.begin()),
-                       std::make_move_iterator(cmake_rewrites.end()));
-
-      append_findings(ScanProjectMigrations(legacy_path.parent_path(), summary.project_name,
-                                            current_version, summary.entrypoint_out_dir));
-    }
+  if (RefreshGeneratedGlue(manifest_path.parent_path(), summary.project_name, current_version,
+                           summary.entrypoint_out_dir)) {
+    REXLOG_INFO("Regenerated generated/rexglue.cmake for SDK v{}", current_version);
   }
 
-  EmitProjectHeader(manifest_path, summary);
-  EmitManualReview(warnings,
-                   fmt::format("Migration: {} site(s) need manual review:", warnings.size()));
-
-  std::vector<OverwriteEntry> consent_view = pre_plan;
-  consent_view.insert(consent_view.end(), post_plan.begin(), post_plan.end());
-  if (auto consent = PromptConsent(consent_view, ctx.skip_upgrade_consent); !consent)
-    return consent;
-
-  if (auto applied = ApplyPlan(pre_plan); !applied)
-    return applied;
-
-  if (auto run_result = RecompileProject(manifest_path, ctx, targets); !run_result) {
-    if (from_legacy) {
-      std::error_code ec;
-      fs::remove(manifest_path, ec);
-      REXLOG_ERROR("Codegen failed; manifest write rolled back. Legacy config is unchanged.");
-    }
+  if (auto run_result = RecompileProject(manifest_path, ctx, targets, ignore_stamp); !run_result) {
     return run_result;
   }
 
-  if (auto applied = ApplyPlan(post_plan); !applied)
-    return applied;
-
   if (!rex::codegen::ManifestConfig::WriteSdkVersionStamp(manifest_path, current_version)) {
-    REXLOG_WARN("Failed to stamp manifest sdkVersion; next run may re-prompt");
+    REXLOG_WARN("Failed to stamp manifest sdkVersion");
   }
   return rex::Ok();
 }
@@ -442,6 +227,8 @@ void RegisterCodegen(CLI::App& parent, const CliContext& ctx, DeferredAction& pe
                   "DLL target to build (repeatable; entrypoint always included)")
       ->type_name("NAME")
       ->take_all();
+  sub->add_flag("--ignore-stamp", args->ignore_stamp,
+                "Regenerate even when inputs match the recorded stamp");
 
   sub->callback([args, &ctx, &pending]() {
     pending = [args, &ctx]() -> Result<void> {
@@ -452,7 +239,7 @@ void RegisterCodegen(CLI::App& parent, const CliContext& ctx, DeferredAction& pe
           return Err<void>(discovered.error());
         path = *discovered;
       }
-      return CodegenFromConfig(path, ctx, args->targets);
+      return CodegenFromConfig(path, ctx, args->targets, args->ignore_stamp);
     };
   });
 }
