+++ src/rexglue/ui/ui.cpp
@@ -12,7 +12,6 @@
 #include "ui.h"
 
 #include <algorithm>
-#include <cctype>
 #include <cstdio>
 #include <iostream>
 #include <ostream>
@@ -36,10 +35,6 @@ void RequireGlobalSink() {
   if (!g_sink) {
     throw std::logic_error("rexglue::ui not initialised; call ui::Init first");
   }
-}
-
-bool StdinIsTty() {
-  return rex::platform::console::is_tty(stdin);
 }
 
 std::string_view LevelLetter(spdlog::level::level_enum lvl) {
@@ -77,20 +72,6 @@ std::string_view LevelColor(spdlog::level::level_enum lvl) {
     default:
       return {};
   }
-}
-
-bool ReadYesAnswer(std::istream& in) {
-  std::string line;
-  if (!std::getline(in, line))
-    return false;
-  std::transform(line.begin(), line.end(), line.begin(),
-                 [](unsigned char c) { return std::tolower(c); });
-  auto first = line.find_first_not_of(" \t\r\n");
-  if (first == std::string::npos)
-    return false;
-  auto last = line.find_last_not_of(" \t\r\n");
-  std::string trimmed = line.substr(first, last - first + 1);
-  return trimmed == "y" || trimmed == "yes";
 }
 
 }  // namespace
@@ -183,20 +164,6 @@ void SetSinkActiveProgress(bool active) {
   g_sink->setActiveProgress(active);
 }
 
-bool ConfirmWithStream(std::string_view question, std::istream& in, bool stdin_is_tty) {
-  {
-    GlobalSinkAccessor acc;
-    acc.writeColored(color::kBold, question);
-    acc.writeRaw(" ");
-    acc.writeColored(color::kDim, "[y/N]:");
-    acc.writeRaw(" ");
-    acc.flush();
-  }
-  if (!stdin_is_tty)
-    return false;
-  return ReadYesAnswer(in);
-}
-
 }  // namespace detail
 
 void Init(const InitOptions& opts) {
@@ -251,42 +218,6 @@ void KeyValueBlock(std::string_view header, std::span<const KeyValueRow> rows) {
   acc.writeRaw("\n");
 }
 
-void PlanTable(std::string_view header, std::span<const PlanRow> rows) {
-  detail::GlobalSinkAccessor acc;
-  if (!header.empty()) {
-    acc.writeColored(color::kBold, header);
-    acc.writeRaw("\n");
-  }
-  std::size_t path_width = 0;
-  for (const auto& r : rows)
-    path_width = std::max(path_width, r.path.size());
-  for (const auto& r : rows) {
-    acc.writeRaw("  [");
-    auto trimmed_action = r.action_label;
-    while (!trimmed_action.empty() && trimmed_action.back() == ' ')
-      trimmed_action.remove_suffix(1);
-    if (trimmed_action == "write") {
-      acc.writeColored(color::kGreen, r.action_label);
-    } else if (trimmed_action == "delete") {
-      acc.writeColored(color::kRed, r.action_label);
-    } else {
-      acc.writeRaw(r.action_label);
-    }
-    acc.writeRaw("] ");
-    acc.writeRaw(r.path);
-    if (!r.reason.empty()) {
-      if (path_width > r.path.size()) {
-        acc.writeRaw(std::string(path_width - r.path.size() + 2, ' '));
-      } else {
-        acc.writeRaw("  ");
-      }
-      acc.writeColored(color::kDim, r.reason);
-    }
-    acc.writeRaw("\n");
-  }
-  acc.writeRaw("\n");
-}
-
 void ManualReviewList(std::string_view header, std::span<const ManualReviewRow> rows) {
   detail::GlobalSinkAccessor acc;
   if (!header.empty()) {
@@ -305,10 +236,6 @@ void ManualReviewList(std::string_view header, std::span<const ManualReviewRow> 
     }
   }
   acc.writeRaw("\n");
-}
-
-bool Confirm(std::string_view question) {
-  return detail::ConfirmWithStream(question, std::cin, StdinIsTty());
 }
 
 void DoneSummary(std::chrono::milliseconds elapsed) {
